+++ LastFMKit/Models/LFMAlbum.h
@@ -55,10 +55,10 @@ NS_SWIFT_NAME(Album)
 @property(strong, nonatomic, readonly) NSString *mbid;
 
 /** The amount of listeners the album has. */
-@property(nonatomic, readonly) unsigned int listeners;
+@property(nonatomic, readonly) NSUInteger listeners;
 
 /** The amount of "scrobbles" the track has. */
-@property(nonatomic, readonly) unsigned int playCount;
+@property(nonatomic, readonly) NSUInteger playCount;
 
 /** An array of all the album's tracks. */
 @property(strong, nonatomic, readonly) NSArray <LFMTrack *> *tracks;
